Миксин AdminRequiredMixin перенесен в core/permissions.py
