+++ adaptive_hockey_federation/analytics/views.py
@@ -2,11 +2,11 @@ from datetime import datetime
 
 from analytics.forms import AnalyticsFilterForm
 from core.constants import GENDER_CHOICES
+from core.permissions import AdminRequiredMixin
 from dateutil.relativedelta import relativedelta
 from django.db.models import Q
 from main.controllers.player_views import PlayersListView
 from main.models import Nosology, Team
-from main.permissions import AdminRequiredMixin
 
 
 class AnalyticsListView(
